SqlTableProvider zrefaktorovan, aby sel upravit na ruzne konkretni implementace SQL, zejmena zpusob urcovani RowNumber pro vkladany radek.

+++ CqrsFramework/Infrastructure/MemoryTableProvider.cs
@@ -149,7 +149,7 @@ namespace CqrsFramework.Infrastructure
             _table.Dispose();
         }
 
-        public long GetMaxRowNumber()
+        public int GetMaxRowNumber()
         {
             return _nextRowNumber - 1;
         }
